Update run.sh
Was getting an error and failing the "Check cert and key file" check.

app_1  | File not found: /letsencrypt/certificates/_.*.xxx.xxx.dev.pem

Don't think it should be using LEGO_DOMAINS for the string replacement as this could be a long line of domain SANs. Replaced with a simple string change to ROUTEROS_DOMAIN if it contains "*" characters.